+++ Product/Helper/Config.php
@@ -4,11 +4,17 @@ namespace Pimgento\Product\Helper;
 
 use \Magento\Framework\App\Helper\AbstractHelper;
 use \Magento\Framework\App\Helper\Context;
+use \Pimgento\Import\Helper\Serializer as Json;
 use \Magento\Store\Model\StoreManagerInterface;
-use \Magento\Framework\Filesystem;
 
 class Config extends AbstractHelper
 {
+    /** Config keys */
+    const CONFIG_PIMGENTO_PRODUCT_ATTR_MAPPING      = 'pimgento/product/attribute_mapping';
+    const CONFIG_PIMGENTO_PRODUCT_CONFIGURABLE_ATTR = 'pimgento/product/configurable_attributes';
+    const CONFIG_PIMGENTO_PRODUCT_TAX_CLASS         = 'pimgento/product/tax_class';
+    const CONFIG_PIMGENTO_PRODUCT_IMAGE_ENABLED     = 'pimgento/image/enabled';
+    const CONFIG_CATALOG_SEO_PRODUCT_URL_SUFFIX     = 'catalog/seo/product_url_suffix';
 
     /**
      * @var \Magento\Store\Model\StoreManagerInterface
@@ -16,15 +22,23 @@ class Config extends AbstractHelper
     protected $_storeManager;
 
     /**
+     * @var Json
+     */
+    protected $serializer;
+
+    /**
      * @param \Magento\Framework\App\Helper\Context $context
      * @param StoreManagerInterface $storeManager
+     * @param Json                  $serializer
      */
     public function __construct(
         Context $context,
-        StoreManagerInterface $storeManager
+        StoreManagerInterface $storeManager,
+        Json $serializer
     )
     {
         $this->_storeManager = $storeManager;
+        $this->serializer = $serializer;
         parent::__construct($context);
     }
 
@@ -35,14 +49,14 @@ class Config extends AbstractHelper
      */
     public function getProductTaxClasses()
     {
-        $classes = $this->scopeConfig->getValue('pimgento/product/tax_class');
+        $classes = $this->scopeConfig->getValue(self::CONFIG_PIMGENTO_PRODUCT_TAX_CLASS);
 
         $result = array();
 
         $stores = $this->_storeManager->getStores(true);
 
         if ($classes) {
-            $classes = unserialize($classes);
+            $classes = $this->serializer->unserialize($classes);
             if (is_array($classes)) {
                 foreach ($classes as $class) {